Fix compilation errors across codebase - Update STT provider integration examples to match current API signatures - Fix Google and Deepgram provider constructor calls in examples - Resolve CDR duration type mismatch (int64 vs float64) - Add missing config imports and fix variable name conflicts - Ensure all core packages and main application compile successfully All changes maintain existing functionality while fixing build issues.

+++ pkg/cdr/service.go
@@ -160,7 +160,7 @@ func (c *CDRService) EndSession(sessionID string, status string, errorMessage *s
 	// Finalize CDR
 	now := time.Now()
 	cdr.EndTime = &now
-	duration := now.Sub(cdr.StartTime).Seconds()
+	duration := int64(now.Sub(cdr.StartTime).Seconds())
 	cdr.Duration = &duration
 	cdr.Status = status
 	cdr.ErrorMessage = errorMessage
